requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Web Scraping & Browser Automation
playwright==1.40.0
requests==2.31.0

# Data Processing & Analysis
pandas==2.1.3
numpy==1.26.2
pdfplumber==0.10.3

# LLM Integration (Google Gemini)
google-generativeai==0.3.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging & Monitoring
python-json-logger==2.0.7

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
